Address review feedback and move all schema-related items into "schema" subdir
