+++ netbox_custom_objects/schema/__init__.py
@@ -1,0 +1,1 @@
+"""Portable COT schema toolkit: format constants, exporter, comparator, executor."""
