+++ netbox_custom_objects/schema/cot_schema_v1.json
